// src/models/user.model.ts
export interface User {

    nome: string;
    quantidade: string;
    start: string;
    end: string;
    id: string;
    categoria: string;
    status: string;
    date: string;
    horarioInicio: string;
    horarioFinal: string;
    local: string;
    inscritos: [];
    email: string;
    descricao: string;
}